revert file rename to `constants.js`
